+++ scratch.js
@@ -12,22 +12,11 @@ const fmt = function (iso) {
   return spacetime(iso).format('{day-short} {nice} {year}')
 }
 
-// let doc = nlp('next weekend')
-let doc = nlp('20th-21st').debug()
-// let num = doc.numbers()
-// num.toNumber()
-// num.toCardinal(false)
-// doc.debug()
-doc.contractions().expand().debug()
+let doc = nlp('1-31')
 // let found = doc.dates({ today: [2016, 7, 13], timezone: 'Canada/Pacific' }).json()[0]
-// console.log(found.date.start)
 // console.log(fmt(found.date.start))
 // console.log(fmt(found.date.end))
 
-// let doc = nlp(`alexandria Daddario`).debug()
-// let doc = nlp(`Paris Berelc`).debug()
-// let doc = nlp(`in alexandria`).debug()
-// doc.debug()
-// doc.sentences().forEach(s => {
-//   s.phrases().debug()
-// })
+// doc.numbers().debug().toCardinal()
+doc.numbers().debug().toOrdinal()
+console.log(doc.text())